src/**/*.csproj: Define CodeAnalysisRuleSet conditionally
This patch may serve to work around a matter of the availability of
the file `AllRules.ruleset`, such that may not exist under some MSBUild
installations, This workaround, as such, may be applied in a manner as
follows, namely in disabling the RunCodeAnalysis property when building
NUnitLite under MSBuild.

~~~~
msbuild NUnitLite-1.0.0/src/framework/nunitlite-4.5.csproj /p:RunCodeAnalysis=false
~~~~

This may serve to prevent a number of warning messages, when building
NUnitLite under MSBuild with some Mono installations.

This patch provides, in effect, a conditional evaluation for the
declaration of the CodeAnalysisRuleSet property in NUnitLite *.csproj
files.